Kernel: Xiaomi kernel changes for Redmi POCO X3 Pro Android R
The Patch based on QualComm release TAG:LA.UM.9.1.r1-07000-SMxxx0.0-1
The kernel config file used is vayu_user_defconfig

+++ disable_dbgfs.sh
@@ -4,26 +4,35 @@
 # disable debugfs for user builds
 export MAKE_ARGS=$@
 
-if [ ${DISABLE_DEBUGFS} == "true" ]; then
-	echo "build variant ${TARGET_BUILD_VARIANT}"
-	if [ ${TARGET_BUILD_VARIANT} == "user" ] && \
-		[ ${ARCH} == "arm64" ]; then
-		echo "combining fragments for user build"
-		(cd $KERNEL_DIR && \
-		ARCH=${ARCH} CROSS_COMPILE=${CROSS_COMPILE}\
-		./scripts/kconfig/merge_config.sh \
-		./arch/${ARCH}/configs/$DEFCONFIG \
-		./arch/${ARCH}/configs/vendor/debugfs.config
-		make ${MAKE_ARGS} ARCH=${ARCH} \
-		CROSS_COMPILE=${CROSS_COMPILE} savedefconfig
-		mv defconfig ./arch/${ARCH}/configs/$DEFCONFIG
-		rm .config)
-	else
-		if [[ ${DEFCONFIG} == *"perf_defconfig" ]] && \
+echo "Enable debugfs baseon ENABLE_DEBUG_FS:'$ENABLE_DEBUG_FS'"
+if [ ${ENABLE_DEBUG_FS} == "true" ]; then
+	echo "Overriding kernel config with CONFIG_DEBUG_FS=y to '$DEFCONFIG'"; \
+	echo -e "\n#KERNEL_CONFIG_OVERRIDE_DEBUG_FS\n" >> ${KERNEL_DIR}/arch/${ARCH}/configs/$DEFCONFIG; \
+	echo "CONFIG_DEBUG_FS=y" >> ${KERNEL_DIR}/arch/${ARCH}/configs/$DEFCONFIG; \
+	echo -e "\n" >> ${KERNEL_DIR}/arch/${ARCH}/configs/$DEFCONFIG;
+else
+
+	if [ ${DISABLE_DEBUGFS} == "true" ]; then
+		echo "build variant ${TARGET_BUILD_VARIANT}"
+		if [ ${TARGET_BUILD_VARIANT} == "user" ] && \
 			[ ${ARCH} == "arm64" ]; then
-			echo "resetting perf defconfig"
-			(cd ${KERNEL_DIR} && \
-			git checkout arch/$ARCH/configs/$DEFCONFIG)
+			echo "combining fragments for user build"
+			(cd $KERNEL_DIR && \
+			ARCH=${ARCH} CROSS_COMPILE=${CROSS_COMPILE}\
+			./scripts/kconfig/merge_config.sh \
+			./arch/${ARCH}/configs/$DEFCONFIG \
+			./arch/${ARCH}/configs/vendor/debugfs.config
+			make ${MAKE_ARGS} ARCH=${ARCH} \
+			CROSS_COMPILE=${CROSS_COMPILE} savedefconfig
+			mv defconfig ./arch/${ARCH}/configs/$DEFCONFIG
+			rm .config)
+		else
+			if [[ ${DEFCONFIG} == *"perf_defconfig" ]] && \
+				[ ${ARCH} == "arm64" ]; then
+				echo "resetting perf defconfig"
+				(cd ${KERNEL_DIR} && \
+				git checkout arch/$ARCH/configs/$DEFCONFIG)
+			fi
 		fi
 	fi
 fi
